### 2024-09-16 #7232 @editorialbot
[PRE REVIEW]: delfies: a Python package for the detection of DNA breakpoints with neo-telomere addition
**Submitting author:** @bricoletc (<a href="http://orcid.org/0000-0002-8921-6005">Brice Letcher</a>)
**Repository:** https://github.com/bricoletc/delfies
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.0
