+++ paper/paper.bib
@@ -77,6 +77,38 @@
   year      = {1887},
 }
 
+@article{Beermann1977,
+author={Beermann, Sigrid},
+title={The diminution of heterochromatic chromosomal segments in Cyclops (Crustacea, Copepoda)},
+journal={Chromosoma},
+year={1977},
+month={Dec},
+day={01},
+volume={60},
+number={4},
+pages={297-344},
+issn={1432-0886},
+doi={10.1007/BF00292858},
+url={https://doi.org/10.1007/BF00292858}
+}
+
+@article{Yu1991,
+  title = {Developmentally programmed healing of chromosomes by telomerase in tetrahymena},
+  volume = {67},
+  ISSN = {0092-8674},
+  url = {http://dx.doi.org/10.1016/0092-8674(91)90077-C},
+  DOI = {10.1016/0092-8674(91)90077-c},
+  number = {4},
+  journal = {Cell},
+  publisher = {Elsevier BV},
+  author = {Yu,  Guo-Liang and Blackburn,  Elizabeth H.},
+  year = {1991},
+  month = nov,
+  pages = {823–832}
+}
+
+
+
 @article{Dedukh2021,
   title = {Delete and survive: strategies of programmed genetic material elimination in eukaryotes},
   volume = {97},
@@ -226,3 +258,5 @@
   year = {2022},
   month = jan 
 }
+
+
