### 2024-09-20 #7232 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again. If some references truly do not have a DOI that is fine, you can leave those unchanged.
- [ ] Can you confirm if your project features (automated) testing? If so, it is good practice to describe and link to testing in the README as well.
- [ ] Can you confirm full coverage of documentation? I recommend more clearly linking to the documentation in the README (e.g. have a section on it).
