### 2024-09-22 e3793e2
Linting: add flake8 config and fix flake8-spotted errors
