### 2024-09-22 6155b31
Updated README
- Table of contents
- Contributing section
