### 2024-09-22 94c05e5
Fix typos in README
