### 2024-09-22 dfe2fae
Added automated test running for releases
