### 2024-09-22 #7232 @bricoletc
Dear @Kevin-Mattheus-Moerman, thanks for these points! They should be addressed below.

> @bricoletc Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
>     * [ ]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again. If some references truly do not have a DOI that is fine, you can leave those unchanged.

I indeed do not think the one reference without a DOI has a DOI, the paper is too old (1887)
>
>     * [ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).
Thank you, I've added a [CONTRIBUTING.md](https://github.com/bricoletc/delfies/blob/main/CONTRIBUTING.md)
>
>     * [ ]  Can you confirm if your project features (automated) testing? If so, it is good practice to describe and link to testing in the README as well.
The project features testing, yes, but it wasn't automated. I have a release github action that is triggered when releasing a new version, and uploads the project to PyPI. Before uploading, I've now added to the action a run through all the [tests](https://github.com/bricoletc/delfies/blob/dfe2fae88564476fd323ac34aaf78b6740833cfe/.github/workflows/release_pypi.yml#L23)

Code coverage of the tests is currently 59%, but all the most important functions are tested. `pytest-cov` report:

```sh
---------- coverage: platform linux, python 3.10.11-final-0 ----------
Name                              Stmts   Miss Branch BrPart  Cover
-------------------------------------------------------------------
delfies/SAM_utils.py                 46      0     16      0   100%
delfies/__init__.py                  13      1      0      0    92%
delfies/breakpoint_foci.py          166     65     64      1    57%
delfies/breakpoint_sequences.py      26      6      8      0    71%
delfies/delfies.py                  104    104     40      0     0%
delfies/interval_utils.py            40      0     11      0   100%
delfies/seq_utils.py                 62      1     28      0    99%
-------------------------------------------------------------------
TOTAL                               457    177    167      1    59%
```
>
>     * [ ]  Can you confirm full coverage of documentation? I recommend more clearly linking to the documentation in the README (e.g. have a section on it).
I'm not sure I fully understood your point, but I've updated the README with a table of contents and added a section pointing to the [detailed documentation](https://github.com/bricoletc/delfies?tab=readme-ov-file#detailed_docs) (I like to keep the main README relatively concise)
