### 2024-10-04 8cbcecb
Fix #1: Update python version support to >=3.9
