+++ pyproject.toml
@@ -10,7 +10,7 @@ readme = "README.md"
 delfies = 'delfies.delfies:main'
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.9.0"
 pysam = "^0.22.0"
 pybedtools = "^0.9.1"
 click = "^8.1.7"
