### 2024-10-14 #7232 @bricoletc
Hey @mahfuz05062 , sure, how about [petaripenev](https://reviewers.joss.theoj.org/reviewers/1517) or [TahiriNadia](https://reviewers.joss.theoj.org/reviewers/1765) . In general I think someone who works in genomics and/or evolution and with expertise in Python would be well-suited.
