### 2024-10-22 9d7d539
Added JOSS status badge to README
