### 2024-10-24 bricoletc/delfies#2 @natir
[JOSS] Paper issue
This issue is related to [JOSS](https://joss.theoj.org/) review process track by issue [#7385](https://github.com/openjournals/joss-reviews/issues/7385). Here I'm going to group my comments on the paper only, another issue will group the other elements.

I add a to-do list to see if my problems are solved or an answer, the fact that it's not completed is not a blocking factor:

- [ ]  It could be intrusting to see memory usage and cpu run time of delfies on different dataset, real or synthetic
- [ ] During reading of paper it's seems MILTEL is the main *competitor* of delfies (I note feature set are different and I could be wrong)  a benchmark on result and memory usage and cpu could interest reader.
- [ ] Could you provides a Sensibility/Specificity metrics of delfies result ?
- [ ] You indicate you test delfies on PacBio Hifi but what is expected behavior of delfies on nanopore, pacbio or illumina
- [ ] Maybe different aligner could have impact on result and performance of delfies could you talk about it
- [ ] Could you talk and show effect of error rate on result of delfies
