### 2024-10-24 bricoletc/delfies#3 @natir
[JOSS] Other comment
This issue is related to [JOSS](https://joss.theoj.org/) review process track by issue https://github.com/openjournals/joss-reviews/issues/7385. Here I'm going to group my comments on code and documentation another issue [group paper comment](https://github.com/bricoletc/delfies/issues/2).

I add a to-do list to see if my concern are solved or an answer, the fact that it's not completed is not a blocking factor:

- [x] Added a complete usage example, including the mapping phase in the readme or documentation. Would make it easier to use delfies for the first time.
- [x] It might be interesting to add functional tests on a small set of synthesized or real data.
- [ ] Add an API documentation building instruction
