### 2024-10-25 bricoletc/delfies#3 @bricoletc
Thanks @natir, will work on these!
