### 2024-11-11 810ae5f
Removed intermediate file writing
Store all the intermediate results in memory instead of in files
