### 2024-11-12 125225e
Add integration test (#3)
- Added a test that simulates reads containing an S2G breakpoint and
  asserts for the breakpoint
- Takes total project coverage from 61% to 74% (`make coverage`)
