+++ tests/test_breakpoint_detection_from_BAM.py
@@ -1,0 +1,152 @@
+from dataclasses import dataclass
+from pathlib import Path
+from random import choice as rand_choice
+from tempfile import NamedTemporaryFile
+
+import pytest
+from pysam import AlignedSegment, AlignmentFile
+from pysam import index as pysam_index
+from pysam import qualitystring_to_array
+
+from delfies import BreakpointType
+from delfies.breakpoint_foci import (
+    BreakpointDetectionParams,
+    find_breakpoint_foci_row_based,
+)
+from delfies.interval_utils import Interval
+from delfies.SAM_utils import DEFAULT_MIN_MAPQ, DEFAULT_READ_FILTER_FLAG
+from delfies.seq_utils import TELOMERE_SEQS, Orientation, randomly_substitute, rev_comp
+
+DEFAULT_NON_TELO_SEQ = "TAACCC"
+DEFAULT_TELO_SEQ = TELOMERE_SEQS["Nematoda"][Orientation.forward]
+DEFAULT_CHROM = "chr1"
+DEFAULT_CHROM_LENGTH = 10e03
+DEFAULT_MIN_TELO_ARRAY_SIZE = 10
+DEFAULT_NUM_TELO_CONTAINING_READS = 150
+EXPECTED_BREAKPOINT_POSITION = 1000
+
+
+@dataclass
+class TeloContainingReadGenerator:
+    read_template: str = DEFAULT_NON_TELO_SEQ * 100 + DEFAULT_TELO_SEQ * 100
+    num_reads: int = 200
+    num_telo_containing_reads: int = DEFAULT_NUM_TELO_CONTAINING_READS
+    telo_start_pos_in_genome: int = EXPECTED_BREAKPOINT_POSITION
+    telo_start_pos_in_read: int = len(DEFAULT_NON_TELO_SEQ) * 100
+    telo_array_size: int = len(DEFAULT_TELO_SEQ) * DEFAULT_MIN_TELO_ARRAY_SIZE * 2
+    num_mutations: int = 0
+    BAM_file: str = None
+
+    def _add_telo_containing_softclipped_sequence(self, read: AlignedSegment) -> None:
+        telo_start_pos = self.telo_start_pos_in_read
+        index_start = rand_choice(range(telo_start_pos))
+        non_telo_seq = self.read_template[index_start:telo_start_pos]
+        telo_seq = self.read_template[
+            telo_start_pos : telo_start_pos + self.telo_array_size
+        ]
+        read.reference_start = self.telo_start_pos_in_genome - len(non_telo_seq)
+        read.cigar = ((0, len(non_telo_seq)), (4, len(telo_seq)))
+        read.query_sequence = randomly_substitute(
+            non_telo_seq, self.num_mutations
+        ) + randomly_substitute(telo_seq, self.num_mutations)
+        read.template_length = len(read.query_sequence)
+        read.query_qualities = qualitystring_to_array("<" * len(read.query_sequence))
+
+    def _add_nontelo_containing_fully_aligned_sequence(
+        self, read: AlignedSegment
+    ) -> None:
+        telo_start_pos = self.telo_start_pos_in_read
+        index_start = rand_choice(range(telo_start_pos))
+        non_telo_seq_1 = self.read_template[index_start:telo_start_pos]
+        non_telo_seq_2 = self.read_template[
+            telo_start_pos - self.telo_array_size : telo_start_pos
+        ]
+        read.reference_start = self.telo_start_pos_in_genome - len(non_telo_seq_1)
+        read.query_sequence = randomly_substitute(
+            non_telo_seq_1, self.num_mutations
+        ) + randomly_substitute(non_telo_seq_2, self.num_mutations)
+        read.cigar = ((0, len(read.query_sequence)),)
+        read.template_length = len(read.query_sequence)
[…]
