### 2024-11-12 6539ef2
Add test coverage reports
- To CONTRIBUTING.md
- To github release action CI
