### 2024-11-12 e7b5f8f
Add codecov badge to README
