### 2024-11-13 d7210e6
Update version and dependencies
- Delfies to v0.7.0
- Datasci dependency to v0.1.1
