+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "delfies"
-version = "0.6.0"
+version = "0.7.0"
 description = "delfies is a tool for the detection of DNA Elimination breakpoints"
 authors = ["Brice Letcher <brice.letcher@ens-lyon.fr>"]
 license = "MIT"
@@ -16,7 +16,7 @@ pybedtools = "^0.9.1"
 click = "^8.1.7"
 rich-click = "^1.7.3"
 pyfastx = "^2.1.0"
-datasci-bricoletc = "^0.1.0"
+datasci-bricoletc = "^0.1.1"
 edlib = "^1.3.9"
 
 [tool.poetry.group.dev.dependencies]
