### 2024-11-13 31ad336
Updated README to address #2
- Added discussion of read aligners
- Added discussion of sequencing technologies
- Added test dataset for running delfies end-to-end (#3)
