### 2024-11-13 b6e2acc
Reformulate README sentences on input data
