+++ paper/paper.md
@@ -11,7 +11,7 @@ authors:
     orcid: 0000-0002-8921-6005
     affiliation: 1 # (Multiple affiliations must be quoted)
 affiliations:
-  - name: Laboratory of Biology and Modeling of the Cell, Ecole Normale Supérieure de Lyon, CNRS UMR 5239, Inserm U1293, University Claude Bernard Lyon 1, Lyon, France
+  - name: Laboratory of Biology and Modelling of the Cell, Ecole Normale Supérieure de Lyon, CNRS UMR 5239, Inserm U1293, University Claude Bernard Lyon 1, Lyon, France
     index: 1
 bibliography: paper.bib
 ---
@@ -101,14 +101,25 @@ For further speed, `delfies` also supports multi-threading.
 `delfies` has already been used to successfully characterise the breakpoints,
 motifs, and retained/eliminated genomes of several nematode genera in the
 family *Rhabditidae*, supporting two upcoming publications (Letcher *et al.*
-and Stevens *et al.*, in preparation). We anticipate this tool can be of broad
-use to researchers studying Programmed DNA Elimination, to characterise species
-known to eliminate but also to discover or screen for elimination in new
-species. This is especially relevant as new long-read and high-coverage
-sequencing data (of both germline and somatic cells) of eukaryotic species
-become increasingly available [@DTOL2022; @EBGP2022]. `delfies` may also be
-useful in other fields of research in which modified chromosomes with
-neo-telomeres are formed and maintained, notably cancer biology.
+and Stevens *et al.*, in preparation). Before these results are made publicly
+available with those publications, the author has prepared a subset of
+publicly-available data from the nematode *Oscheius onirici*, whose elimination
+breakpoint motif has been previously described [@Estrem2023]. The data are
+available on Zenodo [@ZenodoData2024], and consist of a small genome region
+containing a known single breakpoint and alignment files for reads sequenced
+using three distinct technologies (Illumina NovaSeq, Oxford Nanopore
+Technologies PromethION and Pacific Biosciences Sequel II) and spanning
+a range of average read lengths (151bp to 11.9kbp) and per-base qualities (Q11
+to Q28). `delfies` recovers a single, identical breakpoint across all three datasets.
+
+We anticipate this tool can be of broad use to researchers studying Programmed
+DNA Elimination, to characterise species known to eliminate but also to
+discover or screen for elimination in new species. This is especially relevant
+as new long-read and high-coverage sequencing data (of both germline and
+somatic cells) of eukaryotic species become increasingly available [@DTOL2022;
+@EBGP2022]. `delfies` may also be useful in other fields of research in which
+modified chromosomes with neo-telomeres are formed and maintained, notably
+cancer biology.
 
 # Acknowledgements
 
