+++ paper/paper.bib
@@ -260,3 +260,10 @@ url={https://doi.org/10.1007/BF00292858}
 }
 
 
+@dataset{delfies_zenodo_2024,
+  author    = {Brice Letcher},
+  title     = {Dataset supporting the tool 'delfies: a Python package for the detection of DNA breakpoints with neo-telomere addition'},
+  year      = {2024},
+  doi       = {10.5281/zenodo.14101797},
+  publisher = {Zenodo},
+}
