### 2024-11-13 bricoletc/delfies#2 @bricoletc
@natir, thank you for the suggestion of testing `delfies` across the three most common sequencing technologies, I really like this suggestion.

To do this I used a very convenient dataset of Illumina, PacBio and ONT data sequenced on the same species (the nematode *Oscheius onirici*), that I aligned to a small region of the genome with a known breakpoint. I find that `delfies` recovers this breakpoint using all three datasets. I provide these input data and the `delfies` outputs on Zenodo: https://doi.org/10.5281/zenodo.14101797

I now mention this in the README: https://github.com/bricoletc/delfies?tab=readme-ov-file#test_run
And in the paper, here's the new text:

```
+[...] and Stevens *et al.*, in preparation). Before these results are made publicly
+available with those publications, the author has prepared a subset of
+publicly-available data from the nematode *Oscheius onirici*, whose elimination
+breakpoint motif has been previously described [@Estrem2023]. The data are
+available on Zenodo [@ZenodoData2024], and consist of a small genome region
+containing a known single breakpoint and alignment files for reads sequenced
+using three distinct technologies (Illumina NovaSeq, Oxford Nanopore
+Technologies PromethION and Pacific Biosciences Sequel II) and spanning
+a range of average read lengths (151bp to 11.9kbp) and per-base qualities (Q11
+to Q28). `delfies` recovers a single, identical breakpoint across all three datasets.
```

I also now discuss which aligners to use and what they need to report in a new section of the README: https://github.com/bricoletc/delfies?tab=readme-ov-file#input_data

For the more quantitative aspects of your todo list:

- CPU and memory: In the paper I state that on a reasonably high-coverage dataset (85x), `delfies` processes a 250Mbp genome in 2minutes on a single thread. I could add mention of memory usage if you want (it's low), but I think this is 'enough' in the sense of having a very much usable tool
- For a detailed comparison with MILTEL, I decided not to do this and focus on the qualitative differences (stated in the paper). I think this is enough to justify statement of need. I believe it's not the JOSS policy to provide benchmarks of performance against competitors, just that there is an obvious application (https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and discussion of its context in the existing field
- For sensitivity/specificity, I'd repeat the point above, and also adding my 'positive control' on the three sequencing technologies shows that the tool works at a basic level. In upcoming publications (see mentions in the paper), we properly validate identified breakpoints using wet-lab experiments, so I am confident that identified breakpoints are genuine, at least in the nematode species we work on (*Mesorhabditis* genus notably).

Happy to re-revise according to your thoughts on all this!
And on @KatyBrown's when her review comes in
