### 2024-11-13 bricoletc/delfies#3 @natir
Nice work !
