### 2024-11-13 bricoletc/delfies#2 @natir
Nice work !

If I understand correctly, there will be specificity and sensitivity information in a forthcoming paper which will also publish the data required for this type of assessment? Aren't there any data sets already published that would at least give us an approximation of these values?
