### 2024-11-15 dad50bc
Add README section on validating breakpoints (#2)
