### 2024-11-15 bricoletc/delfies#2 @bricoletc
Thanks!

That's not exactly what I meant. What I mean is the best way to validate that a breakpoint is a true, biological breakpoint, is by showing that it is cut in a biological context. We did this for a number of breakpoints identified by delfies in the genus *Mesorhabditis* (~15 or so). But no there will not be any sensitivity/specificity in that paper.

I could do sensitivity/specificity, but it would be on simulated data (there are no big enough sets of biologically-validated breakpoints yet for something like a ROC curve). IMO sensitivity/specificity is overkill considering the journal policy, and also will only be valid for simulated data.

However I can see that for a tool user, it's important to have a sense of whether an identified breakpoint is a true breakpoint or not. So I added  a section called 'Validating breakpoints' in the README:  https://github.com/bricoletc/delfies?tab=readme-ov-file#validating-breakpoints
