### 2024-11-27 #7385 @KatyBrown
Hi @bricoletc, @mahfuz05062, @natir - sorry for my very late review.
Everything looks great to me! I had no problems  installing the software, the documentation is clear and the paper has all the information required.
I only have very minor comments as follows:
- It would be helpful to give the basic command to run the software (`delfies [OPTIONS] GENOME_FNAME BAM_FNAME ODIRN`) in the GitHub README.
- In order to run the test run, I had to index the bam files with samtools.
- The test run code just says "run delfies here" - it doesn't provide instructions on how to do that.
- It would be helpful to provide instructions to run the unit tests

None of these are required for publication but all would be helpful.
Thanks and sorry again for the delay.
