### 2024-12-05 f7f4f70
Updated README to reflect @KatyBrown's review
