### 2024-12-05 #7385 @bricoletc
@KatyBrown, thanks for your feedback, here's my 'rebuttal':

>
    - It would be helpful to give the basic command to run the software (delfies [OPTIONS] GENOME_FNAME BAM_FNAME ODIRN) in the GitHub README.
For this point, I find that the current [Getting started section](https://github.com/bricoletc/delfies#-getting-started) gives this information: the user can either invoke `delfies --help`, or can see that the basic usage is `delfies <genome>.fa.gz <aligned_reads>.bam <output_dir>`. Do you consider this is not enough information?
>
    - In order to run the test run, I had to index the bam files with samtools.
Thank you, I've made an updated version of the test run data that includes `samtools`-indexed BAMs (.bai files). I've updated the link for downloading that new data in the [README](https://github.com/bricoletc/delfies#-test-run-with-real-data)
>
    - The test run code just says "run delfies here" - it doesn't provide instructions on how to do that.
Thank you, I added an example command in the same updated section of the [README](https://github.com/bricoletc/delfies#-test-run-with-real-data)
>
     - It would be helpful to provide instructions to run the unit tests
I agree, but thought this was already present in the [CONTRIBUTING.md file](https://github.com/bricoletc/delfies/blob/main/CONTRIBUTING.md). Do you consider what's in there sufficient? I don't detail exactly how to run the tests using `pytest`, but state how to do this using `make`, do you want me to add a full `pytest` command example?
Also, to be more explicit, I now state in the README contributing section:
```md
Contributions always welcome!

Please see [CONTRIBUTING.md](https://github.com/bricoletc/delfies/blob/main/CONTRIBUTING.md) for how (reporting issues, requesting features, contributing code). This document includes instructions on how to run delfies' unit and functional tests.
```

Also, can you confirm that running `delfies` on the test input worked and produced the expected outputs?
