### 2024-12-19 f14d9e1
Fix typos in paper
