+++ paper/paper.bib
@@ -93,7 +93,7 @@ url={https://doi.org/10.1007/BF00292858}
 }
 
 @article{Yu1991,
-  title = {Developmentally programmed healing of chromosomes by telomerase in tetrahymena},
+  title = {Developmentally programmed healing of chromosomes by telomerase in Tetrahymena},
   volume = {67},
   ISSN = {0092-8674},
   url = {http://dx.doi.org/10.1016/0092-8674(91)90077-C},
