### 2024-12-19 853cea9
Bump version to 0.8.0
