### 2024-12-22 #7385 @mahfuz05062
Hi @bricoletc, I went through the text and it looks good overall. Can you fix one typo: Line 57 - please correct the spelling to "software". Thanks!
