### 2024-12-22 #7385 @mahfuz05062
editorialbot set v0.8.0 as version
