### 2024-12-26 d57c8a4
Fix typo in paper text
