### 2024-12-26 #7385 @bricoletc
> Hi @bricoletc, I went through the text and it looks good overall. Can you fix one typo: Line 57 - please correct the spelling to "software". Thanks!

Hi @mahfuz05062, thanks, I've just corrected this in d57c8a40cd86a7f79699d1892434123992bb80b5
