### 2024-12-26 #7385 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
