### 2025-01-09 #7385 @Kevin-Mattheus-Moerman
@bricoletc
as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:

- [ ] Check `chomosome` which should probably be written `chromosome`
- [ ] Please edit (this is a manually editing process requiring signing in) the Zenodo archive listed version tag to be `v0.8.0` (and be sure to include the `v`).
