### 2025-01-12 #7385 @bricoletc
Dear @Kevin-Mattheus-Moerman,

Thanks, both have been done: paper typo fixed in [latest commit](https://github.com/bricoletc/delfies/commit/6dd330c56e386a076e6963e18f85eb54ff2b7a54) of the software, and v0.8.0 used on [zenodo](https://zenodo.org/records/14526258)
Best
Brice
