### 2025-01-13 #7385 @Kevin-Mattheus-Moerman
@bricoletc please update the paper title to have:

```
title: 'delfies: a Python package for the detection of DNA breakpoints with neo-telomere addition'
```
