### 2025-01-13 8e20801
Added CITATION.cff file
