### 2025-05-05 #7385 @bricoletc
Dear @arfon (cc @Kevin-Mattheus-Moerman), as per my request and our exchanges by email I've added an author on this paper, in commit https://github.com/bricoletc/delfies/commit/7c14fd98e5b36cda31a974a76df13d405da77fb6 on the code repo, in `paper.md`

Can you reprocess/republish the paper accordingly?

Thanks, best,
Brice
